refactor: extract insights sankey card
